pyproject.toml
[tool.poetry]
name = "live-check"
version = "0.1.0"
description = ""
authors = ["duyuxuan <duyuxuan@yixia.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.16"
requests = "^2.32.4"
pydantic-settings = "^2.9.1"
sqlalchemy = "^2.0.41"
pymysql = "^1.1.1"
aioredis = "^2.0.1"
redis = "^6.4.0"
loguru = "^0.7.3"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
rocketmq-client-python = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
